Lagt til ny dokumentasjon for barnetilsyn og flyttet på dokumentfelt til egen fil
