+++ pom.xml
@@ -27,7 +27,7 @@
         <com.amazonaws.version>1.11.820</com.amazonaws.version>
         <token-validation-spring.version>1.2.0</token-validation-spring.version>
         <felles.version>1.20200703124425_7e616f5</felles.version>
-        <kontrakter.version>2.0_20200825152231_6c4a598</kontrakter.version>
+        <kontrakter.version>2.0_20200904102330_d1b3555</kontrakter.version>
         <start-class>no.nav.familie.ef.søknad.ApplicationKt</start-class>
         <!--suppress UnresolvedMavenProperty  Ligger som secret i github-->
         <sonar.projectKey>${SONAR_PROJECTKEY}</sonar.projectKey>
