+++ src/main/kotlin/no/nav/familie/ef/søknad/mapper/kontrakt/DokumentIdentifikator.kt
@@ -25,6 +25,7 @@ enum class DokumentIdentifikator {
     FAKTURA_BARNEPASSORDNING,
     AVTALE_BARNEPASSER,
     ARBEIDSTID,
+    ROTERENDE_ARBEIDSTID,
     TRENGER_MER_PASS_ENN_JEVNALDREDE,
     FOR_SYK_TIL_Å_JOBBE,
